Responding to review feedback, renaming project to adhere to python naming conventions and restructured project
