+++ pyglet_utils/utils.py
