+++ tests/test_game.py
@@ -1,8 +1,6 @@
 import pytest
-import unittest
 
 
-class TestGame(unittest.TestCase):
-    @pytest.mark.skip("placeholder")
-    def test_game(self):
-        assert True
+@pytest.mark.skip("placeholder")
+def test_game(self):
+    assert True
